now using datetime in google charts!
this makes the line chart graphs scale acc to time elapsed, even given
the sometimes irregular intervals at which the app collects data.
previously used strings for dates on the x-axis, which didn't allow for
that.
new setup puts more logic into the template than before, but it's not
too bad