fixed time zone inconsistencies, added pytz in order to fix them
-Execution time data from the citibike-provided json is in New York
time, but it is provided in a string format. Python does not like to
extract time zones from strings, period.
-The GAE data store assumes datetime objects are in UTC, if no time
zone is provided.
-Adding a time zone to a naive datetime object in Python is no simple
matter. That's where pytz comes in.
-The pytz module doesn't play well with GAE though, so for this app I
downloaded the GAE-sanctioned version. You can read about it at
https://code.google.com/p/gae-pytz/. Once unzipped, I copied over the
pytz directory into the root directory of this project. I also had to
unzip the timezoneinfo.zip file that was in the pytz directory. That is
why there are so many files added in this commit. The pytz module would
not work until I copied the directory into this project's root
directory and unzipped timezoneinfo.zip.
-Since this app is only intended to work in the America_New York time
zone, it's possible that it could use pytz without all the extra time
zone info files. Alternatively, this app could work just fine, if a
tzinfo was built for it specifically. Then it wouldn't need the pytz
module.