fixed a mistake on the superlatives page
the results for fewest day passes and fewest week passes were
incorrect. this was because of a klugey way that I'd tried to work
around certain days that had zero of each (because citibike rolled out
annual memberships before day and week passes).
in the end, figured out a way to skip over those days using offset,
instead of filtering and ordering. the Query class requires you order
according to the same attribute the filter is based on, and this wasn't
working for the query that was needed here.